server/pyproject.toml
[project]
name = "Peng-Agent"
version = "2.3.0"
description = "A LLM ongoing chat agent for learning"
authors = [
    { name = "Noah Ding", email = "noah@dingyipeng.com" },
    { name = "Tena Walcott", email = "tw@tenawalcott.com" }
]
dependencies = [
    "anthropic>=0.50.0",
    "bcrypt>=4.2.0",
    "azure-ai-documentintelligence>=1.0.1",
    "azure-identity>=1.21.0",
    "bs4>=0.0.2",
    "fastapi>=0.115.11",
    "google-genai>=1.30.0",
    "importlib>=1.0.4",
    "langchain[openai]>=1.0.3",
    "langchain-core>=1.0.3",
    "langchain-openai>=1.0.2",
    "langchain-qdrant>=1.1.0",
    "langgraph>=0.5.3",
    "boto3>=1.28.0",
    "mysql-connector-python>=9.2.0",
    "numpy>=2.2.3",
    "openai>=1.66.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "pymupdf>=1.25.5",
    "pypdf>=5.3.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.13.3",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
    "wikipedia>=1.4.0",
    "lxml>=5.3.0",
    "email-to>=0.1.0",
    "nest-asyncio>=1.6.0",
    "arize-phoenix-otel>=0.12.1",
    "openinference-instrumentation-langchain>=0.1.49",
    "sqlalchemy>=2.0.39",
    "pymysql>=1.1.1",
    "crawl4ai>=0.7.4",
    "tavily-python>=0.7.11",
    "paramiko>=4.0.0",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-community>=0.4.1",
]
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "config", "models", "handlers", "services"]

[tool.ruff]
extend-exclude = ["archive/**", "test/**"]

